+++ PyTorch/built-in/nlp/M2M-100_for_PyTorch/fairseq/data/indexed_dataset.py
@@ -497,8 +497,8 @@ class MMapIndexedDataset(torch.utils.data.Dataset):
         np_array = np.frombuffer(
             self._bin_buffer, dtype=self._index.dtype, count=size, offset=ptr
         )
-        if self._index.dtype != np.int64:
-            np_array = np_array.astype(np.int64)
+        if self._index.dtype != np.int32:
+            np_array = np_array.astype(np.int32)
 
         return torch.from_numpy(np_array)
